Add basic support for UDPLITE
UDPLITE differs in semantics from UDP only in one particular: the
checksum can be partial rather than full. Aside from that it is
almost an entirely unused protocol, originally intended for
interactive (RTP) applications, and the point in adding
it to netperf was to determine what, if anything could use
it effectively, if it could be routed over the ipv4 or ipv6
internet, etc, etc.

This commit adds basic support for UDPLITE to netperf, adding
UDPLITE_RR and UDPLITE_STREAM tests and does not much change the UDP
code path. It doesn't change the output messages, nor does it have a
specific test for the limited checksum feature in UDPLITE.